+++ src/App.js
@@ -2,7 +2,7 @@
 import './App.css';
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import HomePage from "./pages/HomePage";
-import Add from "./pages/AddQuestion";
+import Add from "./pages/AddQuestionPage/AddQuestion";
 
 
 function App() {
